Deduplikoi juttuja InstructionControllerTestissä
